refactor(sputnik): keep llrt_ prefix
