--- src/sputnik/src/js/apis/node/llrt/buffer/mod.rs
@@ -1,4 +1,0 @@
-pub mod blob;
-mod file;
-
-pub use crate::js::apis::node::llrt::utils::primordials::{BasePrimordials, Primordial};
