+++ src/sputnik/src/js/apis/node/llrt/llrt_buffer/file.rs
@@ -1,6 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-use crate::js::apis::node::llrt::utils::time;
+use crate::js::apis::node::llrt::llrt_utils::time;
 use rquickjs::{
     atom::PredefinedAtom, class::Trace, function::Opt, ArrayBuffer, Coerced, Ctx, Exception,
     IntoJs, Object, Result, Value,
