+++ src/sputnik/src/js/apis/node/llrt/llrt_utils/result.rs
